+++ src/Symfony/Component/Validator/Constraints/MaxCount.php
@@ -18,19 +18,24 @@ use Symfony\Component\Validator\Constraint;
  *
  * @api
  */
-class Range extends Constraint
+class MaxCount extends Constraint
 {
-    public $minMessage = 'This value should be {{ limit }} or more.';
-    public $maxMessage = 'This value should be {{ limit }} or less.';
-    public $invalidMessage = 'This value should be a valid number.';
-    public $min;
-    public $max;
+    public $message = 'This collection should contain {{ limit }} elements or less.';
+    public $limit;
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getDefaultOption()
+    {
+        return 'limit';
+    }
 
     /**
      * {@inheritDoc}
      */
     public function getRequiredOptions()
     {
-        return array('min', 'max');
+        return array('limit');
     }
 }